newSchema relocated and small fixes
