+++ components/news/store.js
@@ -1,8 +1,8 @@
-const db = require('../../models/index')
+const { Entries } = require('../../models/index')
 
 const getAll = async () => {
   try {
-    const news = await db.sequelize.models.Entries.findAll()
+    const news = await Entries.findAll()
     return news
   } catch ({ message: error }) {
     throw new Error(error)
@@ -15,7 +15,7 @@ const addNew = async (oneNew) => {
     const { image } = oneNew
     const { categoryId } = oneNew
     const { type } = oneNew
-    const createdNew = await db.sequelize.models.Entries.create({
+    const createdNew = await Entries.create({
       name,
       content,
       image,
@@ -30,7 +30,7 @@ const addNew = async (oneNew) => {
 
 const getById = async (id) => {
   try {
-    const news = await await db.sequelize.models.Entries.findOne({ where: { id } })
+    const news = await await Entries.findOne({ where: { id } })
     return news
   } catch ({ message: error }) {
     throw new Error(error)
