+++ validate/newsSchema.js
